Expand panel.
