+++ packages/konsole-app/src/containers/index.ts
@@ -3,5 +3,4 @@
 //
 
 export * from './Container';
-export * from './Root';
 export * from './Sidebar';
